♻️ export naming function
