+++ src/app/components/SavedResults/SavedResults.tsx
@@ -2,22 +2,12 @@ import React, { useState } from "react";
 import Modal from "../Modal/Modal";
 import styles from "./SavedResults.module.css";
 import { SearchResults } from "../../../types";
+import namingGitHubData from "../../../utils/namingGitHubData";
 
 function SavedResultsComponent({ searchResult }: SearchResults): JSX.Element {
   const [showModal, setShowModal] = useState<boolean>(false);
   const [showCollapse, setShowCollapse] = useState(false);
-
-  const userName = searchResult.repoName.substring(
-    0,
-    searchResult.repoName.lastIndexOf("/")
-  );
-  const fileName = searchResult.name.substring(
-    0,
-    searchResult.name.lastIndexOf(".")
-  );
-  const repoName = searchResult.repoName.substring(
-    searchResult.repoName.indexOf("/") + 1
-  );
+  const { userName, fileName, repoName } = namingGitHubData(searchResult);
 
   return (
     <div className={styles.collapse}>
