+++ src/app/components/SearchResults/SearchResults.tsx
@@ -2,17 +2,12 @@ import React, { useState } from "react";
 import Modal from "../Modal/Modal";
 import styles from "./SearchResults.module.css";
 import { SearchResults } from "../../../types";
+import namingGitHubData from "../../../utils/namingGitHubData";
 
 function SearchResultsComponent({ searchResult }: SearchResults): JSX.Element {
   const [showModal, setShowModal] = useState<boolean>(false);
 
-  const fileName = searchResult.name.substring(
-    0,
-    searchResult.name.lastIndexOf(".")
-  );
-  const repoName = searchResult.repoName.substring(
-    searchResult.repoName.indexOf("/") + 1
-  );
+  const { fileName, repoName } = namingGitHubData(searchResult);
 
   return (
     <div className={styles.searchResults}>
